Move styled components to their respective files
